+++ lib/warnings/parser/bandit_parser.rb
@@ -5,13 +5,8 @@ module Warnings
   # Parser class for bandit generated json files.
   class BanditParser < Parser
     RESULTS_KEY = 'results'.freeze
-    FILE_TYPES = %i(json).freeze
     NAME = 'Bandit'.freeze
     ERROR_MISSING_KEY = "Missing bandit key '#{RESULTS_KEY}'.".freeze
-
-    def file_types
-      FILE_TYPES
-    end
 
     def parse(file)
       json_hash = json(file)
@@ -33,8 +28,7 @@ module Warnings
       issue.severity = to_severity(hash['issue_severity'])
       issue.message = hash['issue_text']
       issue.line = hash['line_number']
-      issue.category = hash['test_id']
-      issue.name = hash['test_name']
+      issue.category = "#{hash['test_id']}-#{hash['test_name']}"
       @issues << issue
     end