+++ spec/spec_helper/assets.rb
@@ -6,8 +6,10 @@ module Warnings
     BANDIT_JSON = "#{ASSETS_DIR}/bandit.json".freeze
     BANDIT_EMPTY = "#{ASSETS_DIR}/bandit_empty.json".freeze
     BANDIT_MISSING_RESULTS = "#{ASSETS_DIR}/bandit_missing_results.json".freeze
-
     PYLINT_TXT = "#{ASSETS_DIR}/pylint.txt".freeze
+    RUBOCOP_JSON = "#{ASSETS_DIR}/rubocop.json".freeze
+    RUBOCOP_MULTI_JSON = "#{ASSETS_DIR}/rubocop_multi_offenses.json".freeze
+    RUBOCOP_SIMPLE = "#{ASSETS_DIR}/rubocop.txt".freeze
 
     BANDIT_FIRST_ISSUE = {
       code: "2852         except ImportError:\n2853             import pickle\n2854         with open(filename, 'wb') as outf:\n",
@@ -29,6 +31,23 @@ module Warnings
       line: '1',
       category: 'F403',
       message: "'from test_project import *' used; unable to detect undefined names"
+    }.freeze
+
+    RUBOCOP_FIRST_ISSUE = {
+      path: 'spec/lib/danger/danger_core/plugins/dangerfile_gitlab_plugin_spec.rb',
+      offenses: [
+        {
+          severity: 'convention',
+          message: 'Do not use semicolons to terminate expressions.',
+          cop_name: 'Style/Semicolon',
+          corrected: false,
+          location: {
+            line: 82,
+            column: 65,
+            length: 1
+          }
+        }
+      ]
     }.freeze
   end
 end