+++ CHANGELOG.md
@@ -5,11 +5,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.1.0] - 2019-01-23
 ### Added 
-- PylintParser - Support pylint formatted reports
+- PylintParser - Support Pylint formatted reports
+- RuboCopParser - Support RuboCop formatted reports
 
 ### Changed
 - Renamed issue.id to issue.category
+
+### Removed
+- Remove issue.name 
 
 ## [0.0.1] - 2019-01-21
 ### Added